added links
The list of things we need to get started was incomplete in so much as it had links for many of the items listed, but not all of them.   Added links for those items to help totally-new peeps to get started using googles top search results for "best___2021 for coding" for each listed item.